refactor: Organized codebase
- moved commands into their own files.
- moved shared logic into utils.cs
- renamed utils.cs to Utils.cs
